feat: refactor settings and add ens domain tab
