+++ src/components/Settings/SettingsBody/ENSDomainSetting/index.tsx
@@ -1,0 +1,5 @@
+const ENSDomainSetting = () => {
+  return <div>ENSDomainSetting</div>
+}
+
+export default ENSDomainSetting
